created save and fetch methods for addrees, product and customer product entities

+++ src/main/java/edu/goshop_ecommerce/repo/AddressRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.goshop_ecommerce.entity.Address;
 
-public interface AddressRepo extends JpaRepository<Address,Integer>{
+public interface AddressRepo extends JpaRepository<Address,Long>{
